Fixed a big problem that Larry found in our code. Check description.
So, in each class where we need objects from a Thread, we used to do something like:
DriveThread driveThread = new DriveThread("DriveThread");
The problem with this is that that is creating multiple instances of the same Thread.
This is NOT what we want. So instead, we create an instance of the Robot.java class, and in each class where we need objects and stuff from a Thread, we do robot.threadName.whatever. This should work perfectly.